# pkg/blockaction/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a BlockAction API.",
        "title": "BlockAction API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/signin": {
            "post": {
                "description": "會員登入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockAction"
                ],
                "summary": "會員登入",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SigninReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/api.SigninResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/signup": {
            "post": {
                "description": "會員註冊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockAction"
                ],
                "summary": "會員註冊",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SignupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/api.SignupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "conflict",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/personal-info": {
            "get": {
                "description": "會員資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockAction"
                ],
                "summary": "會員資訊",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/api.GetPersonalInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "$ref": "#/definitions/api.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BaseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {}
            }
        },
        "api.GetPersonalInfoResp": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "api.SigninReq": {
            "type": "object",
            "required": [
                "account",
                "password"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "api.SigninResp": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "description": "refresh token",
                    "type": "string"
                },
                "token": {
                    "description": "token",
                    "type": "string"
                }
            }
        },
        "api.SignupReq": {
            "type": "object",
            "required": [
                "account",
                "password",
                "user_name"
            ],
            "properties": {
                "account": {
                    "description": "帳號",
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 6
                },
                "password": {
                    "description": "密碼",
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 8
                },
                "user_name": {
                    "description": "名稱",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                }
            }
        },
        "api.SignupResp": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "name": "BlockAction"
        }
    ]
}